+++ resource-server/src/main/java/com/inhabas/api/web/argumentResolver/ArgumentResolverConfig.java
@@ -1,12 +1,10 @@
-package com.inhabas.api.config;
+package com.inhabas.api.web.argumentResolver;
 
-import com.inhabas.api.web.argumentResolver.LoginMemberArgumentResolver;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
 
 @Configuration
 @RequiredArgsConstructor
