+++ resource-server/src/test/java/com/inhabas/testConfig/InterceptorConfigMockBean.java
@@ -1,6 +1,6 @@
 package com.inhabas.testConfig;
 
-import com.inhabas.api.config.InterceptorConfig;
+import com.inhabas.api.web.interceptor.InterceptorConfig;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
